requirements.txt
h5py
glob
numpy
torch
random
Pillow
shutil
torchvision
